added task submission and task done events in Redis
